requirements.txt
# Core dependencies for WLED Christmas Tree Controller
requests>=2.31.0
websocket-client>=1.6.0
numpy>=1.24.0
pyyaml>=6.0

# Optional: For advanced color operations
pillow>=10.0.0

# Optional: For E1.31/sACN streaming (alternative to DDP)
# sacn>=1.9.0
